organized tests between needarch and needuml more consistant.
