+++ tests/doc_test/doc_needuml_jinja_func_import_negative_tests/index.rst
@@ -1,5 +1,5 @@
-TEST DOCUMENT NEEDARCH JINJA FUNC IMPORT NEGATIV TESTS
-======================================================
+TEST DOCUMENT NEEDUML JINJA FUNC IMPORT NEGATIV TESTS
+=====================================================
 
 .. story:: Story
    :id: US_001
