+++ tests/test_needarch.py
@@ -69,28 +69,6 @@ def test_doc_needarch_jinja_import(test_app):
 
 
 @pytest.mark.parametrize(
-    "test_app",
-    [{"buildername": "html", "srcdir": "doc_test/doc_needarch_jinja_func_import_negative_tests"}],
-    indirect=True,
-)
-def test_doc_needarch_jinja_import_negative(test_app):
-    import subprocess
-
-    app = test_app
-
-    srcdir = Path(app.srcdir)
-    out_dir = srcdir / "_build"
-
-    out = subprocess.run(["sphinx-build", "-M", "html", srcdir, out_dir], capture_output=True)
-
-    assert out.returncode == 1
-    assert (
-        "sphinx_needs.directives.needuml.NeedumlException: "
-        "Jinja function 'import()' is not supported in non-embedded needuml directive." in out.stderr.decode("utf-8")
-    )
-
-
-@pytest.mark.parametrize(
     "test_app", [{"buildername": "html", "srcdir": "doc_test/doc_needarch_jinja_func_need"}], indirect=True
 )
 def test_needarch_jinja_func_need(test_app):
