+++ tests/test_needuml.py
@@ -321,5 +321,27 @@ def test_needuml_jinja_func_need_removed(test_app):
     assert out.returncode == 1
     assert (
         "sphinx_needs.directives.needuml.NeedumlException: "
-        "Jinja function 'import()' is not supported in non-embedded needuml directive." in out.stderr.decode("utf-8")
+        "Jinja function 'need()' is not supported in needuml directive." in out.stderr.decode("utf-8")
+    )
+
+
+@pytest.mark.parametrize(
+    "test_app",
+    [{"buildername": "html", "srcdir": "doc_test/doc_needuml_jinja_func_import_negative_tests"}],
+    indirect=True,
+)
+def test_doc_needarch_jinja_import_negative(test_app):
+    import subprocess
+
+    app = test_app
+
+    srcdir = Path(app.srcdir)
+    out_dir = srcdir / "_build"
+
+    out = subprocess.run(["sphinx-build", "-M", "html", srcdir, out_dir], capture_output=True)
+
+    assert out.returncode == 1
+    assert (
+        "sphinx_needs.directives.needuml.NeedumlException: "
+        "Jinja function 'import()' is not supported in needuml directive." in out.stderr.decode("utf-8")
     )
